requirements.txt
asn1crypto==0.24.0
attrs==18.2.0
atws==0.5.1
cached-property==1.5.1
cchardet==2.1.4
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
cryptography==2.4.2
Cython==0.29.3
dnspython==1.16.0
dnsq==1.1.6
expiringdict==1.1.4
flanker==0.9.0
future==0.17.1
idna==2.8
Jinja2==2.10
MarkupSafe==1.1.0
packaging==19.0
pip-review==1.0
ply==3.11
pycparser==2.19
pyparsing==2.3.1
python-mk-livestatus==0.4
pytz==2018.9
redis==3.0.1
regex==2018.11.22
requests==2.21.0
six==1.12.0
slackclient==1.3.0
suds-jurko==0.6
tld==0.9.2
urllib3==1.24.1
WebOb==1.8.5
websocket-client==0.54.0
